int arg changed to double

+++ include/cephes.h
@@ -36,7 +36,7 @@ extern double ndtri ( double y0 );
 extern double pdtrc ( int k, double m );
 extern double pdtr ( int k, double m );
 extern double pdtri ( int k, double y );
-extern double stdtr ( int k, double t );
+extern double stdtr ( double k, double t );
 extern double stdtri ( double k, double p );
 extern double log1p ( double x );
 extern double expm1 ( double x );
